add profile component

+++ staff/maria-navarrete/full-stack/demo-project/logic/src/deleteUser.js
@@ -1,14 +1,22 @@
-const { models: { User } } = require('data')
+const { models: { User, Note } } = require('data')
 const { validators: { validateId, validatePassword } } = require('commons')
 
 
-function deleteUser(id, password) {
-    validateId(id)
+function deleteUser(userId, password) {
+    validateId(userId)
     validatePassword(password)
-    return User.deleteOne({ id, password })
-        .then(result => {
-            if (result.deletedCount === 0) throw new Error(`wrong user ${id} or password`)
+
+    return Note.deleteMany({ user: userId })
+        .then(() => {
+            return User.deleteOne({ id: userId, password })
+                .then(result => {
+                    if (result.deletedCount === 0) throw new Error(`wrong user ${userId} or password`)
+                    else {
+
+                    }
+                })
         })
-}
+
+    }
 
 module.exports = deleteUser
